refactor: one action selector for both forward replays (P1)
Extracts candidate enumeration, evaluation and tie policy out of the two
hand-mirrored 200-line functions -- dp_battery_algorithm's
_best_action_at_continuous_state and pwl_window_dp's
_pwl_best_action_at_continuous_state -- into core/bess/action_selector.py,
parameterized by a continuation-value evaluator. Both call sites become
thin wrappers that contribute only their eval_V/slope pair.

Zero behavior change, enforced by the bit-parity goldens.
