+++ core/bess/tests/unit/test_idle_tie_break.py
@@ -5,11 +5,11 @@ are within the DP's own value noise, prefer the discharge -- it fails safe
 import numpy as np
 import pytest
 
+from core.bess.action_selector import Candidate, _prefer_load_covering_discharge
 from core.bess.dp_battery_algorithm import (
     SOE_STEP_KWH,
     _best_action_at_continuous_state,
     _discretize_state_action_space,
-    _prefer_load_covering_discharge,
     optimize_battery_schedule,
 )
 from core.bess.pwl_window_dp import (
@@ -20,11 +20,26 @@ from core.bess.settings import BatterySettings
 from core.bess.tests.unit.test_scenarios import build_scenario_inputs
 from core.bess.tie_detection import TIE_NOISE_FACTOR
 
-# Candidate tuple: (value, power, next_soe, new_cost_basis, reward, grid_imported)
-IDLE = (10.00, 0.0, 6.0, 0.0, 0.0, 0.25)
-COVER = (9.995, -1.0, 5.74, 0.0, 0.25, 0.0)  # discharges 1 kW, covers 1 kW net load
-OVER = (9.999, -3.0, 5.21, 0.0, 0.30, 0.0)  # discharges past load -> would export
-CHARGE = (9.99, 2.0, 6.5, 0.0, -0.5, 0.75)
+
+def _candidate(value, power, next_soe, new_cost_basis, reward, grid_imported):
+    """Build a `Candidate` from the positional layout these fixtures were
+    written against, so the cases below stay literally what they were when
+    the tie-breaks lived on 6-tuples (P1 gave them a dataclass)."""
+    return Candidate(
+        power=power,
+        next_soe=next_soe,
+        reward=reward,
+        new_cost_basis=new_cost_basis,
+        grid_imported=grid_imported,
+        value=value,
+    )
+
+
+# Candidate fields: (value, power, next_soe, new_cost_basis, reward, grid_imported)
+IDLE = _candidate(10.00, 0.0, 6.0, 0.0, 0.0, 0.25)
+COVER = _candidate(9.995, -1.0, 5.74, 0.0, 0.25, 0.0)  # covers 1 kW net load
+OVER = _candidate(9.999, -3.0, 5.21, 0.0, 0.30, 0.0)  # past load -> would export
+CHARGE = _candidate(9.99, 2.0, 6.5, 0.0, -0.5, 0.75)
 
 
 def _pick(candidates, best_index, epsilon=0.01, home=0.25, solar=0.0):
@@ -52,7 +67,7 @@ def test_decisive_idle_margin_is_never_swapped():
 
 def test_never_swaps_into_exporting_discharge():
     # Only the over-load discharge is within epsilon -> no eligible swap.
-    candidates = [IDLE, (9.90, -1.0, 5.74, 0.0, 0.25, 0.0), OVER, CHARGE]
+    candidates = [IDLE, _candidate(9.90, -1.0, 5.74, 0.0, 0.25, 0.0), OVER, CHARGE]
     assert _pick(candidates, best_index=0) == 0
 
 
@@ -62,9 +77,9 @@ def test_only_exact_or_under_cover_is_eligible():
     # the action set before this tie-break runs) or a genuine export --
     # never a load-cover swap target, so there is no round-up allowance.
     # balance_zero_p is 1.0 kW (home=0.25 kWh, dt=0.25h).
-    idle = (10.00, 0.0, 6.0, 0.0, 0.0, 0.25)
-    barely_over = (9.999, -1.001, 5.74, 0.0, 0.30, 0.0)
-    exact_cover = (9.995, -1.0, 5.76, 0.0, 0.28, 0.0)
+    idle = _candidate(10.00, 0.0, 6.0, 0.0, 0.0, 0.25)
+    barely_over = _candidate(9.999, -1.001, 5.74, 0.0, 0.30, 0.0)
+    exact_cover = _candidate(9.995, -1.0, 5.76, 0.0, 0.28, 0.0)
 
     assert (
         _pick([idle, barely_over], best_index=0) == 0
@@ -93,7 +108,7 @@ def test_zero_epsilon_is_a_no_op():
 
 
 def test_picks_largest_eligible_coverage_among_ties():
-    partial = (9.997, -0.5, 5.87, 0.0, 0.12, 0.12)  # covers half the load
+    partial = _candidate(9.997, -0.5, 5.87, 0.0, 0.12, 0.12)  # covers half the load
     candidates = [IDLE, partial, COVER]
     # Both discharges are inside epsilon; the fuller cover (1.0 kW) wins.
     assert _pick(candidates, best_index=0) == 2
